+++ app/middlewares/validate-role.js
@@ -1,0 +1,8 @@
+const { UNAUTHORIZED } = require('../constants/errors');
+const { authenticationError } = require('../errors');
+
+module.exports.validateInRole = (userRol) => (req, res, next) => {
+  const { role } = res.locals.user;
+  if (role !== userRol) return next(authenticationError(UNAUTHORIZED));
+  return next();
+};
