+++ test/user-registration.spec.js
